+++ .env
@@ -1,0 +1,2 @@
+MONGO_URI="mongodb+srv://zangwoo:wtKWPTSuLx2dIizc@mongodb-ticku.xjix5.mongodb.net/tickudb?retryWrites=true&w=majority"
+PORT=5000
